+++ client-side(frontend)/lifebird/src/App.css
@@ -1,4 +1,4 @@
-.App {
+/* .App {
   text-align: center;
   
 }
@@ -36,4 +36,4 @@
   to {
     transform: rotate(360deg);
   }
-}
+} */
