+++ client-side(frontend)/lifebird/src/Components/home.css
@@ -1,0 +1,153 @@
+.photo{
+    width:auto;
+    height:100vh; 
+    background-image: url(../../public/home.avif);
+    background-repeat:no-repeat ;
+    background-size: cover; 
+}
+
+.navbar{
+    display: flex;
+    border: 1px solid white;
+    background-color: goldenrod;
+}
+
+.home{
+    margin-left: 730px;
+    margin-top: 20px;
+}
+.doctor{
+    margin-left: 100px;
+    margin-top: 20px;
+}
+
+.lab{
+    margin-left: 100px;
+    margin-top: 20px;
+}
+
+.medicine{
+    margin-left: 100px;
+    margin-top: 20px;
+}
+
+.signup{
+    margin-left: 100px;
+    margin-top: 20px;
+}
+
+.logo{
+    margin-left: 80px;
+    margin-top: 2   0px;
+}
+
+.text{
+    font-size: 40px;
+    margin-left: 100px;
+    color: rgb(21, 21, 23);
+    margin-top: 300px;
+}
+
+.button{
+     height:50px;
+    width:200px; 
+    background-color: goldenrod;
+    margin-left: 100px;
+    border-radius: 5px;
+    border: none;
+    margin-top: 10px;
+}
+
+.left{
+    height:700px;
+    width:50%;
+    border-radius: 10px;
+    margin-top: 10px;
+}
+
+.right{
+    height:700px;
+    width:50%;
+    border-radius: 5px;
+    margin-top: 10px;
+}
+
+.card{
+    display: flex;
+    margin-top: 5px;
+    background-color: rgb(199, 199, 232);
+}
+
+.img{
+    height:350px;
+    width:80%;
+    margin-left: 85px;
+    margin-top: 30px;
+    border-radius: 10px;
+}
+
+.img img{
+    height:100%;
+    width:100%;   
+}
+
+h3{
+    margin-top: 40px;
+}
+
+#learn{
+    height:50px;
+    width:200px;
+    background-color: goldenrod;
+    border-radius: 5px;
+    border: none;   
+}
+.video-layout{
+    width:100%;
+    height:600px;
+    margin-top: 5px;
+}
+.video{
+    height:70%;
+    width:60%;
+    margin-left: 350px;
+    margin-top: 90px;
+}
+.footer{
+    width: 100%;
+    height:400px;
+    border: 1px solid black;
+  
+}
+
+#smallcards div {
+    transition: all 0.5s;
+    border-radius: 10px;
+    box-shadow: rgba(100, 110, 113, 0.838) 0px 2px 8px 0px;
+ }
+ #smallcards div:hover {
+   transform: scale(1.04);
+}
+
+#smallcards div {
+   text-align: center;
+   padding: 20px;
+   margin-left: 40px;
+}
+#smallcards {
+    display: flex;
+    grid-template-columns: repeat(auto-fit, minmax(300px, max-content));
+    gap: 65px;
+    background-color: rgb(215, 219, 241);
+    border-radius: 10px;
+    padding: 20px;
+ }
+ .card3{
+    height:300px;
+    width:300px;
+    border-radius: 10px;
+ }
+ 
+
+
+
