--- client-side(frontend)/lifebird/src/Components1/login.css
@@ -1,76 +1,0 @@
- .out-line{
-    height:500px; 
-    width:500px;  
-    margin-left: 1100px;
-    margin-top: 240px; 
-    
-    background-image: linear-gradient(to right, rgb(11, 138, 242), #aad0d0);  
-   box-shadow: rgba(0, 0, 0, 0.25) 0px 54px 55px, rgba(0, 0, 0, 0.12) 0px -12px 30px, rgba(0, 0, 0, 0.12) 0px 4px 6px, rgba(0, 0, 0, 0.17) 0px 12px 13px, rgba(0, 0, 0, 0.09) 0px -3px 5px;  
-     
- }
-  
- body{
-    background-image: url(../../public/baground.jpg);
-    background-repeat: no-repeat;
-    background-size: cover;
-    height:81vh;
- } 
- .input1{
-    margin: 50px 110px;
- }
-
- 
-  .box{
-    
-    margin: 0 auto;
-    margin-top: 150px;
-    width: 243px;
-    
- }  
- 
- .details1{
-    width: 100%;
-    font-size: 20px;
-    outline: none;
-    background-color: rgba(0, 0, 0 , 0);
-    border: none;
-    border-bottom: 1px solid #FFF;
-    color: white;
- }
- img{
-    height:200px;
-    width:200px;
-    border-radius: 201px;
- }
- 
- .logos img{
-    height:50px;
-    width:50px;border-radius: 50px;
-    margin-top: 50px;
- }
- .logos{
-    display: flex;
- }
- .facebook{
-    margin-left: 150px;
- }
- .insta{
-    margin-left: 20px;
- }
- .twitter{
-    margin-left: 20px;
- }
- .login{
-	height:40px;
-	width:150px;
-	border-radius: 5px;
-	background-color: rgb(14, 184, 241);
-	
-}
-#head1{
-   margin-top: 60px;
-}
- 
- 
- 
-
